chore: refactor specific integration extensions out of the integration file
